Fix workflow file
Specify exact version 10.22.0. This ensures that the minimum version
in package.json is always tested by CI, as well as the latest version
(10.x). For example, if a new feature is added in 10.23 and we use it,
we want CI to fail so that we know to update the minimum in
package.json.